// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/ratetagallowancetimemodel.ts
export class RateTagAllowanceTime {
    id: string;
    branchId: string;
    allowanceRateTagForId: string;
    maxTime: number;
    minTime: number;
    activeFrom: any;
    activeTo: any;
    isArchived: boolean;
    timeStamp: any;
    isBankHoliday: boolean;
    allowanceRateTagForIds: any = [];
}